refactor(ui): change the main struct (again)
